LPS-61466 refactor SearchEngineUtil into module
